modules.xml: Update to 42.alpha
Pin gnome-bluetooth versions, in preparation of gnome-bluetooth3. Add
gnome-text-editor and gtksourceview5. Add new Vala API version for
rawhide. Pin librest to 0.8 on rawhide, to ignore 1.0 with new
soversion. Keep libgit2-glib at 1.0 on rawhide only for now. Add
libgweather4, and ensure that it and the old gweather are kept at the
right versions for each branch. Update librsvg2 to 2.53 for rawhide.