Replace fnmatch globbing with version limits
fnmatch is just not good enough for what we need. For example, we want
to update Fedora 41 to the latest gcr 4.3 version, but not update to
4.4. So I wrote:

  <project name="gcr">
    <release version="f41">4.3.*</release>
    <release version="f42">4.4.*</release>
    <release version="f43">4.4.*</release>
  </project>

But this is broken because there exists a 4.3.91 version. A human
immediately knows that this should be treated as a 4.4 version and we
should not update Fedora 41 to this verison, but there's no way to tell
this to fnmatch.

To fix this, I import some version comparison code from GNOME releng and
switch the entire modules.xml to use version limits instead of version
globs, exactly like GNOME releng's tarball-conversion.xml. So now we
specify limits like this:

  <project>
    <release version="f41">4.3.90</release>
    <release version="f42">4.4.90</release>
    <release version="f43">4.4.90</release>
  </project>

Every limit needs to be one version higher than the max version you want
to update to.