Imply --no-mockbuild when simulating
mclazy is designed to use in two stages: First we run a simulation to
see if there's packages that need an update, and what we'd update the
packages to. After verifying that the updates all make sense, we pass
`--no-simulate` to run the real update

For simulating we should prioritize for speed, so we should default to
--no-simulate. But when pushing, we should optimize for correctness and
so running the mockbuild will prevent us from pushing a broken commit to
dist-git.

Fixes https://github.com/mcatanzaro/mclazy/issues/6